#1 - Additional expandsion of crypto.c into crypto_config and crypto_user files;
